+++ test/fonts/lexText.test.ts
@@ -119,7 +119,7 @@ test("should support speed code tokens", () => {
 
 test("should support font tokens", () => {
   expect(
-    lexText("OldFont!F:0c7c1705-dab6-476e-a5ce-17de23b3d591!NewFont")
+    lexText("OldFont!F:0c7c1705-dab6-476e-a5ce-17de23b3d591!NewFont"),
   ).toEqual([
     {
       type: "text",
@@ -170,6 +170,96 @@ test("should preserve newlines after wait code tokens", () => {
     {
       type: "text",
       value: "\nAfter",
+    },
+  ]);
+});
+
+test("should provide empty previewValue for unmatched octal codes", () => {
+  expect(lexText("Before\\003After")).toEqual([
+    {
+      type: "text",
+      value: "Before",
+    },
+    {
+      type: "text",
+      value: "\\003",
+      previewValue: "",
+    },
+    {
+      type: "text",
+      value: "After",
+    },
+  ]);
+});
+
+test("should provide escaped previewValue for octal escape codes", () => {
+  expect(lexText("Before\\005\\012After")).toEqual([
+    {
+      type: "text",
+      value: "Before",
+    },
+    {
+      type: "text",
+      value: "\\005\\012",
+      previewValue: "\n",
+    },
+    {
+      type: "text",
+      value: "After",
+    },
+  ]);
+});
+
+test("should provide escaped previewValue for octal escape codes + non octal escaped character", () => {
+  expect(lexText("Before\\005CAfter")).toEqual([
+    {
+      type: "text",
+      value: "Before",
+    },
+    {
+      type: "text",
+      value: "\\005C",
+      previewValue: "C",
+    },
+    {
+      type: "text",
+      value: "After",
+    },
+  ]);
+});
+
+test("should provide previewValue for octal codes outside of control code range", () => {
+  expect(lexText('Before\\042"After')).toEqual([
+    {
+      type: "text",
+      value: "Before",
+    },
+    {
+      type: "text",
+      value: "\\042",
+      previewValue: '"',
+    },
+    {
+      type: "text",
+      value: '"After',
+    },
+  ]);
+});
+
+test("should provide previewValue for characters with codes beyond 128", () => {
+  expect(lexText("Before\\251After")).toEqual([
+    {
+      type: "text",
+      value: "Before",
+    },
+    {
+      type: "text",
+      value: "\\251",
+      previewValue: "©",
+    },
+    {
+      type: "text",
+      value: "After",
     },
   ]);
 });